Implement icon as custom view
I want to be able to drag and drop folders onto the icon to add them
to the list of watched directories; to do that, the icon needs to be
a custom view.

Right now, this breaks showing the menu when clicking on the icon,
but I'll work on fixing that.