Banish singletons from Squash
Singletons are an ugly anti-pattern and I hate them. Originally, I
conceived that it may be prudent to allow different types of
*Manager classes to be created (perhaps pointing to different storage
locations, for example), and then have a "default one". At any rate,
it sort of made sense to only have one instance of a *Manager object.

But I realize that was foolhardy. For one thing, it creates really
tight coupling between the *Manager classes and SQApplication; if
SQApplication _really_ only needs one canonical instance (or several),
then it should create them as-needed and store them as instance
variables. For another thing, singletons are just plain ugly.

But all that seems overengineered, so just ditch them entirely.